pyproject.toml
[tool.poetry]
name = "onboarding-project"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "onboarding_project"}]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.1.1"
djangorestframework = "^3.14.0"
django-phonenumber-field = "^7.0.0"
phonenumbers = "^8.12.56"
PyYAML = "^6.0"
django-rest-swagger = "^2.2.0"
drf-yasg = "1.20.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
